backend: Reuse actor proxy to self
We now create an actor proxy to self in on_start() and reuse this
object when creating the pyspotify session and when starting playback.

As before, we still create two proxies to SpotifyBackend at startup:
one in Mopidy core and one in Mopidy-Spotify to refer to itself.

The difference from before is that we now create one actor proxy less
when starting playback.