search: Use the new Web API for search
Spotify has discontinued their Metadata API, which libspotify uses. This
means that search in libspotify, and thus pyspotify is not working. This
changes the search to use the new Web API which is supported.

This is a first step for fixing search, and makes it work again. We
probably want to further improve this, and possibly move calling the API
into pyspotify. Note that some parts of the Web API require OAuth (i.e.
different auth than libspotify requires), but luckily search doesn't
require any auth.

The search results from the Web API misses some data compared to what
libspotify does. The fields that are missing is bitrate and date from
tracks, and artists and date from albums. It is possible to additional
requests to get the artists and release date for albums, but this is
left for a later improvement.

Additionally, the search results will now return unavailable tracks as
well. The API supports limiting the results to only available tracks by
setting the query parameter `market` to either a two level country code
or from_token (if you're authenticated). However, I'm not sure if it's
possible to get the users market from libspotify and we don't do
authenticated requests, so I've not included this yet.

Fixes #89.