fix: Implement search limits correctly for Web API
The Web API only has a single limit, which limits the number of items of
each type. This sets that limit to the max of the config options
search_album_count, search_artist_count and search_track_count.
Furthermore it limits the results for each type to its respective config
limit to mimic the old behavior.

The max value for these config options has been decreased to 50, because
this is the max value that the Web API allows.