Playlists: Improve performance of subsequent playlist track lookups.
Even once we have all playlist data from the Web API, future track lookups are done
through libspotify which must then load the data again from scratch. Clients may try
to lookup all tracks in a playlist and this is therefore now very slow for large
playlists. Previously, libspotify would start loading playlist track data in the
background on first access to the playlist_container; this enabled any subsequent track
lookups we did to be fast. Since libspotify will do this background loading for any
object once it has a Link to it, we can recreate this optimisation by simply grabbing
libspotfy track Links for all playlist track URIs once we get them from the Web API.
We do need to hold a reference to the Links else the Libspotify Link will be freed
and the data will not be loaded.