tests: fix compatibility with pykka 2.0.
When asserting mock.call in list(some_call), the some_call ends up on the lhs.
In our failing test cases, the some_call on the lhs includes an ActorProxy and the mock.call on the rhs includes the mock.ANY, so we get ActorProxy.__eq__(mock.ANY), which returns False.
In Pykka 1.x there was no ActorProxy.__eq__() defined so we ended up doing mock.ANY.__eq__(ActorProxy), which returned True.